% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Api.R
\name{check_version}
\alias{check_version}
\title{check if server and binding versions are compatible}
\usage{
check_version(user_key, url = "https://api.rosette.com/rest/v1/")
}
\arguments{
\item{user_key}{- Rosette API authentication key}

\item{url}{- url for Rosette Api}
}
\value{
Returns the response from the Rosette API checkVersion endpoint
}
\description{
check if server and binding versions are compatible
}

