% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Api.R
\name{post_endpoint}
\alias{post_endpoint}
\title{POST request to specified Rosette API endpoint}
\usage{
post_endpoint(userKey, parameters, endpoint, custom_headers, url)
}
\arguments{
\item{userKey}{- Rosette API authentication key}

\item{parameters}{- parameters list to be passed to specified Rosette API endpoint}

\item{endpoint}{- Rosette API endpoint to be utilized}

\item{custom_headers}{- custom headers for Rosette Api}

\item{url}{- url for Rosette Api}
}
\value{
Returns the response from the Rosette API
}
\description{
POST request to specified Rosette API endpoint
}

