% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Api.R
\name{get_endpoint}
\alias{get_endpoint}
\title{GET request to specified Rosette API endpoint}
\usage{
get_endpoint(user_key, endpoint, customHeaders, url)
}
\arguments{
\item{user_key}{- Rosette API authentication key}

\item{endpoint}{- Rosette API endpoint to be utilized}

\item{customHeaders}{- custom headers for Rosette Api}

\item{url}{- url for Rosette Api}
}
\value{
Returns the response from the Rosette API
}
\description{
GET request to specified Rosette API endpoint
}

