% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Api.R
\name{mutipart_call}
\alias{mutipart_call}
\title{make a multipart POST request}
\usage{
mutipart_call(user_key, parameters, endpoint, customHeaders, url)
}
\arguments{
\item{user_key}{- Rosette API authentication key}

\item{parameters}{- parameters list to be passed to specified Rosette API endpoint}

\item{endpoint}{- Rosette API endpoint to be utilized}

\item{customHeaders}{- custom headers for Rosette Api}

\item{url}{- url for Rosette Api}
}
\value{
Returns the response from the Rosette API
}
\description{
make a multipart POST request
}

